+++ app/src/main/java/com/elfennani/aniwatch/models/Episode.kt
@@ -4,7 +4,7 @@ data class Episode(
     val id: String,
     val allanimeId: String,
     val animeId: Int,
-    val episode: Int,
+    val episode: Double,
     val name: String,
     val dubbed: Boolean,
     val thumbnail: String?,
